better support for C++. fix for SYSCLK_FREQ on STM32F4 to get more precise results (e.g. 48000000 not 47999996). i2c engine reanimated, improved and runs on STM32F4. Support for I2C3. timer support much improved, now there's a tim_init() function just like for other peripherials. libuseful/clike.h now can easily and peacefully co-exist with libc, or work without it using "native" function names (e.g. sprintf not _sprintf). printf: a couple fixes for formatting with character-padding.

+++ include/useful/useful-generic.h
@@ -153,13 +153,13 @@ DECLARE_CMPXCHG (u64, uint64_t)
 #endif
 
 #if __UINTPTR_MAX__ == __UINT16_MAX__
-#  define cmpxchg_p(v,e,d) cmpxchg_u16((uintptr_t *)v, (uintptr_t *)e, (uintptr_t)d)
+#  define cmpxchg_p(v,e,d) cmpxchg_u16((uint16_t *)v, (uint16_t *)e, (uint16_t)d)
 #elif __UINTPTR_MAX__ == __UINT32_MAX__
-#  define cmpxchg_p(v,e,d) cmpxchg_u32((uintptr_t *)v, (uintptr_t *)e, (uintptr_t)d)
+#  define cmpxchg_p(v,e,d) cmpxchg_u32((uint32_t *)v, (uint32_t *)e, (uint32_t)d)
 #elif __UINTPTR_MAX__ == __UINT64_MAX__
-#  define cmpxchg_p(v,e,d) cmpxchg_u64((uintptr_t *)v, (uintptr_t *)e, (uintptr_t)d)
+#  define cmpxchg_p(v,e,d) cmpxchg_u64((uint64_t *)v, (uint64_t *)e, (uint64_t)d)
 #else
-#  error "Unknown size of uintptr_t type"
+#  error "Unknown pointer type size"
 #endif
 
 // -------------------------------------------------------------------------- //
